[TL42-103] feat: Profile 컴포넌트 제작 완료
기존에 UserInfo라는 컴포넌트로 작업했으나 이전에 Profile이라는 이름으로
작명해 뒀던것을 맞추기 위해서 ProfileContent, Profile컴포넌트로 이름을
바꾸었음.
