+++ front-end/src/UI/Pages/Profile/index.tsx
@@ -1,5 +1,17 @@
 import React from 'react';
+import { ModalBody, ModalCloseButton } from '@chakra-ui/react';
+import RoutedModal from '../../Templates/RoutedModal';
+import ProfileContent from '../../Templates/ProfileContent';
 
-export function Profile() {
-  // return ();
+function Profile() {
+  return (
+    <RoutedModal>
+      <ModalCloseButton />
+      <ModalBody>
+        <ProfileContent isMyself />
+      </ModalBody>
+    </RoutedModal>
+  );
 }
+
+export default Profile;
