--- front-end/src/UI/Templates/ProfileHeader/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export function ProfileHeader() {
-  // return ();
-}
